+++ main_test.go
@@ -1,57 +1,38 @@
-package go_testify_yPracticum
+package main
 
 import (
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"net/http"
 	"net/http/httptest"
-	"strconv"
 	"strings"
 	"testing"
 
-	_ "github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-var cafeList = map[string][]string{
-	"moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
-}
-
-func mainHandle(w http.ResponseWriter, req *http.Request) {
-	countStr := req.URL.Query().Get("count")
-	if countStr == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("count missing"))
-		return
-	}
-
-	count, err := strconv.Atoi(countStr)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong count value"))
-		return
-	}
-
-	city := req.URL.Query().Get("city")
-
-	cafe, ok := cafeList[city]
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong city value"))
-		return
-	}
-
-	if count > len(cafe) {
-		count = len(cafe)
-	}
-
-	answer := strings.Join(cafe[:count], ",")
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(answer))
-}
-
 func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	// Тестируем count > totalCount
 	totalCount := 4
+	req := httptest.NewRequest("GET", "/cafe?count=6&city=moscow", nil)
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	list := strings.Split(responseRecorder.Body.String(), ",")
+	assert.Len(t, list, totalCount)
+}
+func TestMainHandlerWhenCityIsNotSupported(t *testing.T) {
+	// Тестируем город в параметре `city` не поддерживается
+	req := httptest.NewRequest("GET", "/cafe?count=2&city=london", nil)
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	assert.Equal(t, "wrong city value", responseRecorder.Body.String())
+}
+
+func TestMainHandlerWhenOkAndBodyNotEmpty(t *testing.T) {
 	// Тестируем код ответа ОК, тело ответа не пустое
 	req := httptest.NewRequest("GET", "/cafe?count=5&city=moscow", nil) // здесь нужно создать запрос к сервису
 	responseRecorder := httptest.NewRecorder()
@@ -61,19 +42,4 @@ func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
 	assert.Equal(t, http.StatusOK, responseRecorder.Code)
 	require.NotEmpty(t, responseRecorder.Body.String())
 
-	// Тестируем город в параметре `city` не поддерживается
-	req = httptest.NewRequest("GET", "/cafe?count=2&city=london", nil)
-	responseRecorder = httptest.NewRecorder()
-	handler.ServeHTTP(responseRecorder, req)
-
-	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
-	assert.Equal(t, "wrong city value", responseRecorder.Body.String())
-
-	// Тестируем count > totalCount
-	req = httptest.NewRequest("GET", "/cafe?count=6&city=moscow", nil)
-	responseRecorder = httptest.NewRecorder()
-	handler.ServeHTTP(responseRecorder, req)
-
-	list := strings.Split(responseRecorder.Body.String(), ",")
-	assert.Len(t, list, totalCount)
 }
